refactor: modularize GM runtime by extracting entrance, mechanic_hint, faction_travel, updates, and rng into separate modules
